backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
qdrant-client==1.12.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
openai==1.58.1
langchain==0.3.13
langchain-openai==0.2.12
llama-index==0.12.1
pgvector==0.3.1
psycopg2-binary==2.9.10
redis==5.2.1
httpx==0.28.1
tenacity==9.0.0
opentelemetry-api==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-exporter-prometheus==0.49b1
bcrypt==4.2.1
PyJWT==2.10.1
python-multipart==0.0.20
pydantic==2.10.4